Trait tower_fault::decider::Decider

pub trait Decider<R> {
    fn decide(&self, req: &R) -> bool;
}

Trait for deciding if a fault should be injected for a given request or response.

Required methods

fn decide(&self, req: &R) -> bool

Decide if a fault should be injected for a given request or response.