Module tower_fault::latency

Latency injection for tower

Layer that injects latency randomly into a service.

Usage

use tower_fault::latency::LatencyLayer;
use tower::{service_fn, ServiceBuilder};

// Initialize a LatencyLayer with a 10% probability of injecting
// 200 to 500 milliseconds of latency.
let latency_layer = LatencyLayer::new(0.1, 200..500);

let service = ServiceBuilder::new()
    .layer(latency_layer)
    .service(service_fn(my_service));

Decider

The decider is used to determine if a latency should be injected or not. This can be a boolean, float, Bernoulli distribution, a closure, or a custom implementation of the Decider trait.

For more information, see the decider module.

use tower_fault::latency::LatencyLayer;

// Never inject latency.
LatencyLayer::new(false, 200..500);
// Always inject 200-500 ms of latency.
LatencyLayer::new(true, 200..500);

// Inject latency 30% of the time.
LatencyLayer::new(0.3, 200..500);

// Inject latency based on the request content.
LatencyLayer::new(|req: &MyRequest| req.value % 2 == 0, 200..500);

Distribution

The latency distribution is used to determine the duration of the latency injected in the service. The distribution can be a Range, RangeInclusive, static value, a closure, or a custom implementation of the Distribution trait.

use tower_fault::latency::LatencyLayer;

// Latency between 200 and 500 milliseconds.
LatencyLayer::new(0.3, 200..500);
LatencyLayer::new(0.3, 200..=500);

// Latency between 200 and 500 milliseconds, using floats for
// extra precision.
LatencyLayer::new(0.3, 200.0..500.0);

// Fixed latency of 300 milliseconds.
LatencyLayer::new(0.3, 300);

// Closure that returns a latency based on the request content.
LatencyLayer::new(0.3, |req: &MyRequest| req.value);

Structs

LatencyLayer

Layer that randomly adds latency to the service.

LatencyService

Service that randomly injects latency into a service.

Traits

Distribution

Trait that returns a random latency.